Remove AbstractPropertyNode and inline its logic to RuntimeBeanNode
