--- subprojects/model-core/src/main/java/org/gradle/internal/properties/bean/AbstractPropertyNode.java
@@ -1,59 +1,0 @@
-/*
- * Copyright 2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.internal.properties.bean;
-
-import com.google.common.base.Equivalence;
-
-import javax.annotation.Nullable;
-
-public abstract class AbstractPropertyNode<T> {
-    private final String propertyName;
-    private final AbstractPropertyNode<T> parentNode;
-
-    public AbstractPropertyNode(@Nullable AbstractPropertyNode<T> parentNode, @Nullable String propertyName) {
-        this.propertyName = propertyName;
-        this.parentNode = parentNode;
-    }
-
-    protected String getQualifiedPropertyName(String childPropertyName) {
-        return propertyName == null ? childPropertyName : propertyName + "." + childPropertyName;
-    }
-
-    @Nullable
-    public String getPropertyName() {
-        return propertyName;
-    }
-
-    @Nullable
-    protected AbstractPropertyNode<T> findNodeCreatingCycle(T childValue, Equivalence<? super T> nodeEquivalence) {
-        if (nodeEquivalence.equivalent(getNodeValue(), childValue)) {
-            return this;
-        }
-        if (parentNode == null) {
-            return null;
-        }
-        return parentNode.findNodeCreatingCycle(childValue, nodeEquivalence);
-    }
-
-    abstract protected T getNodeValue();
-
-    @Override
-    public String toString() {
-        //noinspection ConstantConditions
-        return propertyName == null ? "<root>" : getPropertyName();
-    }
-}
