+++ subprojects/model-core/src/main/java/org/gradle/internal/properties/bean/RuntimeBeanNode.java
@@ -24,22 +24,20 @@ import org.gradle.internal.reflect.validation.TypeValidationContext;
 import javax.annotation.Nullable;
 import java.util.Queue;
 
-public abstract class RuntimeBeanNode<T> extends AbstractPropertyNode<Object> {
+public abstract class RuntimeBeanNode<T> {
 
     private final T bean;
+    private final String propertyName;
+    private final RuntimeBeanNode<T> parentNode;
 
-    protected RuntimeBeanNode(@Nullable RuntimeBeanNode<?> parentNode, @Nullable String propertyName, T bean) {
-        super(parentNode, propertyName);
+    protected RuntimeBeanNode(@Nullable RuntimeBeanNode<T> parentNode, @Nullable String propertyName, T bean) {
+        this.propertyName = propertyName;
+        this.parentNode = parentNode;
         this.bean = Preconditions.checkNotNull(bean, "Null is not allowed as nested property '%s'", propertyName);
     }
 
     public T getBean() {
         return bean;
-    }
-
-    @Override
-    protected Object getNodeValue() {
-        return getBean();
     }
 
     public abstract void visitNode(PropertyVisitor visitor, Queue<RuntimeBeanNode<?>> queue, RuntimeBeanNodeFactory nodeFactory, TypeValidationContext validationContext);
@@ -50,12 +48,37 @@ public abstract class RuntimeBeanNode<T> extends AbstractPropertyNode<Object> {
         return nodeFactory.create(this, qualifiedPropertyName, bean);
     }
 
-    public void checkCycles(String propertyName, Object childBean) {
-        AbstractPropertyNode<?> nodeCreatingCycle = findNodeCreatingCycle(childBean, Equivalence.identity());
+    @Nullable
+    public String getPropertyName() {
+        return propertyName;
+    }
+
+    protected String getQualifiedPropertyName(String childPropertyName) {
+        return propertyName == null ? childPropertyName : propertyName + "." + childPropertyName;
+    }
+
+    public void checkCycles(String propertyName, T childBean) {
+        RuntimeBeanNode<?> nodeCreatingCycle = findNodeCreatingCycle(childBean, Equivalence.identity());
         Preconditions.checkState(
             nodeCreatingCycle == null,
             "Cycles between nested beans are not allowed. Cycle detected between: '%s' and '%s'.",
             nodeCreatingCycle, propertyName);
     }
-}
 
+    @Nullable
+    private RuntimeBeanNode<T> findNodeCreatingCycle(T childValue, Equivalence<? super T> nodeEquivalence) {
+        if (nodeEquivalence.equivalent(getBean(), childValue)) {
+            return this;
+        }
+        if (parentNode == null) {
+            return null;
+        }
+        return parentNode.findNodeCreatingCycle(childValue, nodeEquivalence);
+    }
+
+    @Override
+    public String toString() {
+        //noinspection ConstantConditions
+        return propertyName == null ? "<root>" : getPropertyName();
+    }
+}
